+++ package/kedro_viz/data_access/repositories/__init__.py
@@ -1,4 +1,5 @@
-"""`kedro_viz.data_access.repositories` defines repositories to centralise access to application data."""
+"""`kedro_viz.data_access.repositories` defines repositories to
+centralise access to application data."""
 from .catalog import CatalogRepository
 from .graph import GraphEdgesRepository, GraphNodesRepository
 from .modular_pipelines import ModularPipelinesRepository
