+++ package/kedro_viz/data_access/repositories/graph.py
@@ -1,4 +1,5 @@
-"""`kedro_viz.data_access.repositories.graph` defines interface to centralise access to graph objects."""
+"""`kedro_viz.data_access.repositories.graph` defines interface to
+centralise access to graph objects."""
 # pylint: disable=missing-class-docstring,missing-function-docstring
 from typing import Dict, Generator, List, Optional, Set
