+++ package/kedro_viz/data_access/repositories/tags.py
@@ -1,4 +1,5 @@
-"""`kedro_viz.data_access.repositories.tags` defines repository to centralise access to tags data."""
+"""`kedro_viz.data_access.repositories.tags` defines repository to
+centralise access to tags data."""
 # pylint: disable=missing-class-docstring,missing-function-docstring
 from typing import Iterable, List, Set
