+++ package/kedro_viz/server.py
@@ -34,11 +34,9 @@ def populate_data(
     if creatinge an api app from project.
     """
     if session_store_location:
-        database_engine, sessionmaker_instance = create_db_engine(
-            session_store_location
-        )
+        database_engine, session_class = create_db_engine(session_store_location)
         Base.metadata.create_all(bind=database_engine)
-        data_access_manager.set_db_session(sessionmaker_instance)
+        data_access_manager.set_db_session(session_class)
 
     data_access_manager.add_catalog(catalog)
     data_access_manager.add_pipelines(pipelines)
