+++ package/tests/test_data_access/test_repositories/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+
+
+def identity_func(x):
+    return x
+
+
+@pytest.fixture
+def identity():
+    yield identity_func
