+++ package/tests/test_server.py
@@ -31,6 +31,11 @@ def patched_create_api_app_from_file(mocker):
     yield mocker.patch("kedro_viz.api.apps.create_api_app_from_file")
 
 
+@pytest.fixture
+def example_session_store_location(tmp_path):
+    yield tmp_path / "session_store.db"
+
+
 @pytest.fixture(autouse=True)
 def patched_load_data(
     mocker, example_catalog, example_pipelines, example_session_store_location
@@ -53,7 +58,6 @@ class TestServer:
         patched_uvicorn_run,
         example_catalog,
         example_pipelines,
-        example_session_store_location,
     ):
         run_server()
         # assert that when running server, data are added correctly to the data access manager
